% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_internal.R
\name{breaks_calc}
\alias{breaks_calc}
\title{Calculate break points for density intervals}
\usage{
breaks_calc(x)
}
\description{
Break points are at: 2%, 5%, 10%, 15%, 20%, 25%, 30%, 35%, 40%
}
\keyword{internal}
