% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_internal.R
\name{data_sort}
\alias{data_sort}
\title{Sort data.frame}
\usage{
data_sort(x, col1 = 1, col2 = NA)
}
\arguments{
\item{x}{data.frame to sort}

\item{col1}{integer; the primary column to sort by}

\item{col2}{integer; the secondary column to sort by}
}
\value{
Sorted data.frame object
}
\description{
Sort a data.frame object by the values in one or two columns
}
\keyword{internal}
