% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_create}
\alias{ensemble_create}
\title{Create ensemble of SDM predictions}
\usage{
ensemble_create(x, x.pred.idx, y, y.abund = NULL)
}
\arguments{
\item{x}{is a list of sf objects;
all sf objects must have the same geometry}

\item{x.pred.idx}{vector of names or column indices of predicted density column in
each element of x; must be the same length as x}

\item{y}{rescaling method;
one of: "abundance", "normalization", "standardization", or "sumto1"}

\item{y.abund}{value to which to rescale abundances of x;
ignored if y is not "abundance"}
}
\description{
Description
}
