% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSDM-package.R
\docType{package}
\name{eSDM-package}
\alias{eSDM-package}
\alias{eSDM}
\title{eSDM: A tool for creating and exploring ensembles of predictions from species-distribution models}
\description{
Provides functions + gui (Shiny app)
}
\keyword{package}
