% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSDM_GUI.R
\name{eSDM_GUI}
\alias{eSDM_GUI}
\title{eSDM GUI}
\source{
https://swoodman.shinyapps.io/eSDM/
}
\usage{
eSDM_GUI()
}
\description{
The eSDM graphical user interface (a shiny app) for creating ensembles
  using predictions from SDMs
}
