% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_prep.R
\name{eval_prediction}
\alias{eval_prediction}
\title{Returns ROCR::prediction() output of density and sightings columns
Helper function 2 for AUC and TSS calcs}
\usage{
eval_prediction(x, y, dens.idx, sight.idx, overlap.out = NA)
}
\arguments{
\item{x}{an object of class sf; validation data}

\item{y}{an object of class sf; SDM predictions}

\item{dens.idx}{name or index of column in \code{x} with density data}

\item{sight.idx}{name or index of column in \code{x} with sight data}

\item{overlap.out}{output of \code{eSDM::eval_overlap)()}; if NA then
\code{eSDM::eval_overlap(x, y, dens.idx)} is called within this function}
}
\value{
ROCR::prediction() output of density and sightings columns
}
\description{
Returns ROCR::prediction() output of density and sightings columns
Helper function 2 for AUC and TSS calcs
}
