% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{eval_rmse}
\alias{eval_rmse}
\title{RMSE}
\usage{
eval_rmse(pres.pts, abs.pts, model.data, model.dens.idx, prediction.out = NA)
}
\value{
RMSE for given presence numbers and model predictions
}
\description{
RMSE
}
