% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{eval_tss}
\alias{eval_tss}
\title{Calculates TSS for given presence and absence points and model predictions
Also calculates cutoff value for predictions that maximizes sens + spec}
\usage{
eval_tss(x, y, dens.idx, sight.idx, prediction.out = NA)
}
\arguments{
\item{x}{an object of class sf; validation data}

\item{y}{an object of class sf; SDM predictions}

\item{dens.idx}{name or index of column in \code{y} with density data}

\item{sight.idx}{name or index of column in \code{x} with sight data}

\item{prediction.out}{output of \code{eSDM::eval_prediction()}; if NA then
\code{eSDM::eval_prediction(x, y, dens.idx, sight.idx)} is called within
this function}
}
\value{
Returns TSS value (can return vector also with cutoff value)
}
\description{
Calculates TSS for given presence and absence points and model predictions
Also calculates cutoff value for predictions that maximizes sens + spec
}
