% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{evaluation_metrics}
\alias{evaluation_metrics}
\title{Evaluation metrics}
\usage{
evaluation_metrics(x, y, x.idx, y.idx, count.flag = FALSE)
}
\arguments{
\item{x}{SDM predictions; an object of class sf}

\item{y}{validation data; an object of class sf}

\item{x.idx}{name or index of column in \code{x} with
prediction values}

\item{y.idx}{name or index of column in \code{y} with validation data;
this data may be count or presence/absence data depending on \code{count.flag}}

\item{count.flag}{logical indicating whether the data in column \code{y.idx}
is count data (\code{TRUE}) or presence/absence (\code{FALSE})}
}
\value{
A numeric vector with AUC, TSS and RMSE values, respectively.
  If \code{count.flag == FALSE}, the RMSE value will be \code{NA}
}
\description{
Calculate AUC, TSS, and RMSE for given density predictions and validation data
}
\details{
If \code{count.flag == TRUE}, then \code{eSDM::model_abundance(x, x.idx, FALSE)} will be run
  to get predicted abundance in order to calculate RMSE.
  Note that this assumes the data in column \code{x.idx} of \code{x} are density values.
  If \code{count.flag == FALSE}, all of the values in column \code{x.idx} of \code{x} must be \code{0} or \code{1}
}
\examples{
evaluation_metrics(preds.1, validation.data, 1, "sight")

evaluation_metrics(preds.1, validation.data, 1, "count", TRUE)

}
