% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gshhg.l.L16}
\alias{gshhg.l.L16}
\title{GSHHG world map}
\format{A sfc object with class c("sfc_GEOMETRY", "sfc"):
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.soest.hawaii.edu/pwessel/gshhg/}
}
\usage{
gshhg.l.L16
}
\description{
Description, low resolution.
}
\keyword{datasets}
