% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_abundance.R
\name{model_abundance}
\alias{model_abundance}
\title{Calculate predicted abundance}
\usage{
model_abundance(x, dens.idx, sum.abund = TRUE)
}
\arguments{
\item{x}{object of class sf; SDM with density predictions}

\item{dens.idx}{name or index of column in \code{x} with density predictions.
Can be a vector of names or indices}

\item{sum.abund}{logical indicating whether to return the total predicted
abundnace (TRUE) or the abundance for for each prediction polygon (FALSE)}
}
\value{
Either a single number (total predicted abundnace for the SDM) or
  a numerical vector with the predicted abundance for
  each prediction polygon
}
\description{
Calculates the predicted abundance by multiplying the values in the
  column(s) with density predictions by area of their prediciton polygon
  (calcualted using sf::st_area())
}
