% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vector_processing.R
\name{models.rescale}
\alias{models.rescale}
\title{## Rescale spdf.orig columns cols.data to data in new.abundances}
\usage{
models.rescale(spdf.list, abund.new)
}
\description{
## Rescale spdf.orig columns cols.data to data in new.abundances
}
