% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSDM-internal.R
\name{mround}
\alias{mround}
\title{Round 'x' to nearest 'base' value}
\usage{
mround(x, base, floor.use = FALSE, ceiling.use = FALSE)
}
\description{
Round 'x' to nearest 'base' value
}
\keyword{internal}
