% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_plots_funcs.R
\name{multiplot}
\alias{multiplot}
\title{Title}
\usage{
multiplot(models.toplot, data.name, plot.titles, perc.num, col.num, col.pal,
  leg.inc, leg.labels)
}
\arguments{
\item{models.toplot}{List of sf objects to plot}
}
\description{
Produce thing for plotting
}
