% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{multiplot_layout}
\alias{multiplot_layout}
\title{title}
\usage{
multiplot_layout(models.toplot, data.names, plot.titles, perc.num, col.pal,
  leg.labels, plot.ncol, plot.nrow, axis.cex.curr, main.cex.curr,
  leg.width = 3)
}
\arguments{
\item{models.toplot}{list of sf objects}
}
\description{
General function for plotting sf object
}
