% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vector_processing.R
\name{na_pred_message}
\alias{na_pred_message}
\title{## Generate message reporting length of x
This message was built to refer to prediction values}
\usage{
na_pred_message(x)
}
\description{
## Generate message reporting length of x
This message was built to refer to prediction values
}
