% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_internal.R
\name{na_weight_message}
\alias{na_weight_message}
\title{Generate message reporting length of x}
\usage{
na_weight_message(x, y)
}
\description{
This message was built to refer to weight values, including if any non-NA
  prediction values corresponded to NA weight values
}
\keyword{internal}
