% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vector_processing.R
\name{na_which}
\alias{na_which}
\title{Determine which elements of the x are
one of "N/A", "n/a", "na", "NaN", or ""}
\usage{
na_which(x)
}
\description{
Determine which elements of the x are
one of "N/A", "n/a", "na", "NaN", or ""
}
