% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_plots_funcs.R
\name{plot.multi.display}
\alias{plot.multi.display}
\title{Title}
\usage{
\method{plot}{multi.display}(list.data.display)
}
\arguments{
\item{list.data.display}{List of proper things}
}
\description{
Produce gtable for single- or multi-plot for display
}
