% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_plots_funcs.R
\name{plot.multi.download}
\alias{plot.multi.download}
\title{Title}
\usage{
\method{plot}{multi.download}(list.data.download)
}
\arguments{
\item{list.data.display}{List of proper things}
}
\description{
Produce gtable for single- or multi-plot for download
}
