% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_plots_funcs.R
\name{plot_multi_download}
\alias{plot_multi_download}
\title{Title}
\usage{
plot_multi_download(list.data.download)
}
\arguments{
\item{list.data.display}{List of proper things}
}
\description{
Produce thing for download
}
