% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preds.1}
\alias{preds.1}
\title{Sample SDM density predictions}
\format{An object with class c("data.frame", "sf")}
\usage{
preds.1
}
\description{
Sample SDM density predictions; created by importing
  Sample_predictions_2.csv into the GUI, exporting predictions, and then
  clipping them to the SoCal_bite.csv region
}
\keyword{datasets}
