% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preds}
\alias{preds}
\alias{preds.1}
\alias{preds.2}
\alias{preds.3}
\title{Sample SDM density predictions}
\format{
Objects of class \code{sf} with a column of density predictions (name: \code{Density}) and
  a simple feature list column (name: \code{geometry}).
  \code{preds.1} also has a second column of sample density predictions (name: \code{Density2}),
  as well as \code{Var1} and \code{Var2}, representing the variance

  \code{preds1}: An object of class sf (inherits from data.frame) with 325 rows and 5 columns.

  \code{preds2}: An object of class sf (inherits from data.frame) with 1891 rows and 2 columns.

  \code{preds3}: An object of class sf (inherits from data.frame) with 1445 rows and 2 columns.






}
\usage{
preds.1

preds.2

preds.3
}
\description{
\code{preds.1}, \code{preds.2}, and \code{preds.3} are objects of class \code{\link[sf]{sf}} that serve as
  sample sets of SDM density predictions for the \code{eSDM} package
}
\details{
\code{preds.1} sample SDM density predictions created by importing
  Sample_predictions_2.csv into the eSDM GUI, exporting predictions, and then
  clipping them to the SoCal_bite.csv region.
  Also manually added two variance columns (numbers are randomly generated with a max of 0.01)

\code{preds.2} sample SDM density predictions created by importing
  Sample_predictions_1.csv into the eSDM GUI, exporting predictions, and then
  clipping them to the SoCal_bite.csv region

\code{preds.3} is a set of sample SDM density predictions created by importing
  Sample_predictions_4_gdb into the eSDM GUI, exporting predictions, and then
  clipping them to the SoCal_bite.csv region
}
\keyword{datasets}
