% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_plots_funcs.R
\name{preview.ll}
\alias{preview.ll}
\title{title}
\usage{
preview.ll(sf.ll, data.name, title.ll, perc, axis.cex, main.cex)
}
\arguments{
\item{sf.ll}{sf object in crs.ll}
}
\description{
General function for plotting sf object
}
