% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_interactive}
\alias{preview_interactive}
\title{title}
\usage{
preview_interactive(sdm.ll, data.name, perc, col.pal, leg.labels = NULL,
  title.ll = NULL, leg.title = NULL)
}
\arguments{
\item{sdm.ll}{sf object in crs.ll}
}
\description{
General function for plotting sf object
}
