% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_to_sfc.R
\name{pts_to_sfc_centroids}
\alias{pts_to_sfc_centroids}
\title{Create polygons from centroid coordinates}
\usage{
pts_to_sfc_centroids(x, y, crs.prov = NULL)
}
\arguments{
\item{x}{data frame with at least two columns;
the first two columns must contain longitude and latitude coordinates, respectively.
All other columns are ignored}

\item{y}{numeric; shortest distance from centroid to any of the sides,
and thus also half the length of one side of the polygon}

\item{crs.prov}{crs value for polygon being created;
value must be readable by \code{\link[sf]{st_crs}}}
}
\value{
Object of class \code{sfc} of type \code{POLYGON} with the provided crs
}
\description{
Create polygon(s) from a data frame with coordinates of the polygon centroid(s)
}
\details{
This function was designed for someone who reads in a .csv file
  with a grid of coordinates representing SDM prediction points and needs to create
  non-overlapping prediction polygons with the .csv file coordinates as the centroids.
  However, it can be used to create square polygons around any points, even if those polygons overlap.

  The created polygons are oriented so that, in a 2D plane, their edges are parallel to either the x or the y axis.

  If \code{crs.prov} is not specified, then the \code{crs} attribute of the polygon(s) will be \code{NULL}.
}
\examples{
x <- data.frame(
  lon = c(5, 10, 15, 20, 5, 10, 15, 20),
  lat = c(rep(5, 4), rep(10, 4))
)
pts_to_sfc_centroids(x, 2.5, 4326)

# Combine data and created sfc object to make sf object
x <- data.frame(
  lon = c(5, 10, 15, 20, 5, 10, 15, 20),
  lat = c(5, 5, 5, 5, 10, 10, 10, 10),
  sdm.pred = runif(8),
  sdm.pred2 = runif(8)
)
x.sfc <- pts_to_sfc_centroids(x, 2.5, 4326)
x.sf <- sf::st_sf(x[, 3:4], x.sfc, agr = "constant")

}
