% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_to_sfc.R
\name{pts_to_sfc_coords}
\alias{pts_to_sfc_coords}
\title{Create sfc_polygon object from data.frame with polygon coordinates}
\usage{
pts_to_sfc_coords(x, crs.prov = NULL)
}
\arguments{
\item{x}{data.frame with two columns; the columns must contain longitude
and latitude coordinates, respectively}

\item{crs.prov}{crs value of new sfc_polygon object;
value must be interpretable by st_crs(), see st_crs() documentation}
}
\description{
Create sfc_polygon object from data.frame with only longitude and
  latitude columns, respectively, which contain the polygon coordinates
}
\examples{
x <- data.frame(
  lon = c(40, 40, 50, 50, 40, NA, 20, 20, 30, 30, 20),
  lat = c(0, 10, 10, 0, 0, NA, 0, 10, 10, 0, 0)
)
pts_to_sfc_coords(x, 4326)

}
