% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_to_sfc.R
\name{pts_to_sfc_grid}
\alias{pts_to_sfc_grid}
\title{Create sfc_polygon object from data.frame with grid coordinates}
\usage{
pts_to_sfc_grid(x)
}
\arguments{
\item{x}{thing}
}
\description{
Create sfc_polygon object from data.frame with only longitude and
  latitude columns, respectively, which contain the coordinates
  of the center of each polygon of a grid
}
\examples{


}
