% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_to_sfc.R
\name{pts_to_sfc_vertices}
\alias{pts_to_sfc_vertices}
\title{Create polygons from vertice coordinates}
\usage{
pts_to_sfc_vertices(x, crs.prov = NULL)
}
\arguments{
\item{x}{data frame with at least two columns;
the first two columns must contain longitude and latitude coordinates, respectively.
All other columns are ignored}

\item{crs.prov}{crs value for polygon being created;
value must be readable by \code{\link[sf]{st_crs}}}
}
\value{
Object of class \code{sfc} of type \code{POLYGON} with the provided crs
}
\description{
Create polygon(s) from a data frame with the coordinates of the polygon vertices
}
\details{
Vertices of different polygons must be demarcated by rows with values of \code{NA}
  in both the first and second columns (i.e. the longitude and latitude columns).

  If \code{crs.prov} is not specified, then the \code{crs} attribute of the polygon(s) will be \code{NULL}.
}
\examples{
x <- data.frame(
  lon = c(40, 40, 50, 50, 40, NA, 20, 20, 30, 30, 20),
  lat = c(0, 10, 10, 0, 0, NA, 0, 10, 10, 0, 0)
)
pts_to_sfc_vertices(x, 4326)

}
