% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_read.R
\name{shiny.read.shp}
\alias{shiny.read.shp}
\title{Read GIS shpaefile from Shiny fileInput}
\source{
\url{https://github.com/leonawicz/nwtapp/blob/master/mod_shpPoly.R}
}
\usage{
shiny.read.shp(file.in.list)
}
\arguments{
\item{file.in.list}{The list returned by shiny::fileInput()}
}
\value{
A sf object
}
\description{
Read in a GIS shapefile from a fileInput ouptput in a Shiny app
}
\examples{

}
