% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vector_processing.R
\name{substr_right}
\alias{substr_right}
\title{## Get last n element from string x
From https://stackoverflow.com/questions/7963898}
\usage{
substr_right(x, n)
}
\description{
## Get last n element from string x
From https://stackoverflow.com/questions/7963898
}
