% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vector_processing.R
\name{substrRight}
\alias{substrRight}
\title{## Get last n element from string x
From https://stackoverflow.com/questions/7963898}
\usage{
substrRight(x, n)
}
\description{
## Get last n element from string x
From https://stackoverflow.com/questions/7963898
}
