% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validation.data}
\alias{validation.data}
\title{Sample validation data points}
\format{An object with class c("data.frame", "sf")}
\usage{
validation.data
}
\description{
Sample SDM predictions; created from Validation_data.csv clipped to the
  SoCal_bite.csv region
}
\keyword{datasets}
