% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vector_processing.R
\name{zero.range}
\alias{zero.range}
\title{### Determine whether all values in x are equal, aka have a zero range
From Hadley Wickham}
\usage{
zero.range(x, tol = .Machine$double.eps^0.5)
}
\description{
### Determine whether all values in x are equal, aka have a zero range
From Hadley Wickham
}
