% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_internal.R
\name{zero_range}
\alias{zero_range}
\title{Determine whether all values in x are equal}
\usage{
zero_range(x, tol = .Machine$double.eps^0.5)
}
\description{
From Hadley Wickham on stack overflow
}
\keyword{internal}
