% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vector_processing.R
\name{data_sort}
\alias{data_sort}
\title{Base Frequencies}
\usage{
data_sort(x, col1 = 1, col2 = NA)
}
\arguments{
\item{col1}{integer; the primary column to sort by}

\item{col2}{integer; the secondary column to sort by}
}
\value{
Sorted data.frame
}
\description{
Load csv file from given shiny file input
}
\examples{


}
