% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_create}
\alias{ensemble_create}
\title{Create ensemble of SDM predictions}
\usage{
ensemble_create(x, x.idx, y = NULL)
}
\arguments{
\item{x}{object of class \code{sf}}

\item{x.idx}{vector of column names or numerical indices;
indicates which columns in \code{x} will be used to create the ensemble}

\item{y}{weights for the ensemble; either a numeric vector the same length as \code{x} or
a data frame (or tibble) with the same number of rows as \code{x} and \code{ncol(y) == length(x.idx)}.
If y is a numeric vector, its values (i.e. the weights) must sum to 1.
The default value is \code{1 / length(x.idx)}, i.e. an unweighted ensemble}
}
\value{
object of class \code{sf} with two columns: 'Pred.ens' and 'geometry';
  'Pred_ens' constists of the ensemble predictions and 'geometry' is the simple feature geometry list-column.
  The \code{sf} object attribute \code{agr} is set as 'constant' for 'Pred.ens'
}
\description{
Create a weighted or unweighted ensemble of SDM predictions
}
\details{
\code{ensemble_create} is intended to be used after overlaying predictions with \code{\link{overlay_sdm}} and
  (if desired) rescaling the overlaid predictions with \code{\link{ensemble_rescale}}.

  \code{ensemble_create} includes functionality for ensembling methods provided in \link{eSDM_GUI},
  although not for regional weighting', which currently must be done manually if not using the GUI.
  For instance, if \code{y.weights} is a data frame, then the 'Pixel-level spatial weights' ensembling method is performed.
}
\examples{
ensemble_create(preds.1, c("Density", "Density2"))
ensemble_create(preds.1, c(1, 2), c(0.2, 0.8))

weights.df <- data.frame(runif(325), c(rep(NA, 100), runif(225)))
ensemble_create(preds.1, c("Density", "Density2"), weights.df)

}
