% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_rescale}
\alias{ensemble_rescale}
\title{Rescale SDM predictions}
\usage{
ensemble_rescale(x, x.pred.idx, y, y.abund = NULL)
}
\arguments{
\item{x}{is a list of sf objects;
all sf objects must have the same number of rows}

\item{x.pred.idx}{names or column indices of predicted density column in
each element of x; must be the same length as x}

\item{y}{rescaling method;
one of: "abundance", "normalization", "standardization", or "sumto1"}

\item{y.abund}{value to which to resclae abundances of x;
ignored if y is not "abundance"}
}
\description{
Rescale specified columns in a list of SDM prediction
}
