% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_rescale}
\alias{ensemble_rescale}
\title{Rescale SDM predictions}
\usage{
ensemble_rescale(x, x.idx, y, y.abund = NULL)
}
\arguments{
\item{x}{object of class \code{sf}}

\item{x.idx}{vector of column names or column indices;
indicates which columns in \code{x} will be rescaled}

\item{y}{rescaling method; must be either "abundance" or "sumto1".
See 'Details' section for descriptions of the rescaling methods}

\item{y.abund}{numeric value; ignored if \code{y} is not \code{"abundance"}}
}
\value{
The \code{sf} object \code{x} with the columns specified by \code{x.idx} rescaled
}
\description{
Rescale specified columns for each element in a list of SDM predictions
}
\details{
\code{ensemble_rescale} is intended to be used after overlaying predictions with \code{\link{overlay_sdm}} and
  before creating ensembles with \code{\link{ensemble_create}}.
  The provided rescaling methods are as follows:
  \describe{
  \item{'abundance' - Rescale the density values so that the predicted abundance is \code{y.abund}}{}
  \item{'sumto1' - Rescale the density values so their sum is 1}{}
  }
}
\examples{
ensemble_rescale(preds.1, c("Density", "Density2"), "abundance", 50)
ensemble_rescale(preds.1, c(1, 2), "sumto1")

}
