% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSDM_gui.R
\name{eSDM_gui}
\alias{eSDM_gui}
\title{eSDM GUI}
\source{
https://swoodman.shinyapps.io/eSDM/
}
\usage{
eSDM_gui()
}
\description{
The eSDM graphical user interface (a shiny app) for creating ensembles
  using predictions from SDMs
}
