% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esdm_gui.R
\name{esdm_gui}
\alias{esdm_gui}
\title{Open the eSDM GUI}
\usage{
esdm_gui(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Logical with default of \code{TRUE}; passed to \code{launch.browser} argument of \code{\link[shiny]{runApp}}}
}
\description{
Open the eSDM graphical user interface (GUI); an R Shiny app for creating ensemble predictions using SDM predictions.
  The GUI can also be accessed at \url{https://swoodman.shinyapps.io/eSDM/}
}
