% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSDM_GUI.R
\name{eSDM_GUI}
\alias{eSDM_GUI}
\title{eSDM GUI}
\usage{
eSDM_GUI()
}
\description{
The eSDM graphical user interface (GUI); an R Shiny app for creating ensemble predictions using SDM predictions.
  The GUI can also be accessed at \url{https://swoodman.shinyapps.io/eSDM/}
}
