% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSDM_GUI.R
\name{eSDM_GUI}
\alias{eSDM_GUI}
\title{eSDM GUI}
\usage{
eSDM_GUI()
}
\description{
The eSDM graphical user interface (GUI): a Shiny app for creating ensemble
  predictions using predictions from SDMs. You can also access the gui at
  https://swoodman.shinyapps.io/eSDM/
}
