% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSDM_GUI.R
\name{eSDM_GUI}
\alias{eSDM_GUI}
\title{Open the eSDM GUI}
\usage{
eSDM_GUI(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Logical with default of \code{TRUE}; passed to \code{launch.browser}
argument of \code{\link[shiny]{runApp}}}
}
\description{
Open the eSDM graphical user interface (GUI);
  an R Shiny app for creating ensemble predictions using SDM predictions.
}
\seealso{
\url{https://smwoodman.github.io/eSDM/}
}
