% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{eval_auc}
\alias{eval_auc}
\title{Functions for calculating metric values}
\usage{
eval_auc(pres.pts, abs.pts, model.data, model.dens.idx, plot.name = "",
  prediction.out = NA)
}
\value{
AUC for given p/a points and model predictions
}
\description{
Functions for calculating metric values
}
