% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{eval_auc}
\alias{eval_auc}
\title{Functions for calculating metric values}
\usage{
eval_auc(x, y, dens.idx, sight.idx, prediction.out = NA, plot.name = NA)
}
\arguments{
\item{x}{an object of class sf; validation data}

\item{y}{an object of class sf; SDM predictions}

\item{dens.idx}{name or index of column in \code{y} with density data}

\item{sight.idx}{name or index of column in \code{x} with sight data}

\item{prediction.out}{output of \code{eSDM::eval_prediction()}; if NA then
\code{eSDM::eval_prediction(x, y, dens.idx, sight.idx)} is called within
this function}

\item{plot.name}{the title of the ROCR plot; if \code{NA} then
the ROCR plot is not printed}
}
\value{
AUC value for given validation data and SDM predictions
}
\description{
Functions for calculating metric values
}
