% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{eval_overlap}
\alias{eval_overlap}
\title{Helper functions for calculating model evaluation metrics}
\usage{
eval_overlap(pres.pts, abs.pts, model.data, model.dens.idx)
}
\description{
Return df with overlap of presence/absence data and model
Returned df contains data columns for density, abundance, ...
...sightings (p/a)) flag, and number of animals observed
Helper function 1 for AUC, TSS, and RMSE calcs
}
