% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_prep.R
\name{eval_overlap}
\alias{eval_overlap}
\title{Helper functions for calculating model evaluation metrics}
\usage{
eval_overlap(x, y, dens.idx = NA)
}
\arguments{
\item{x}{an object of class sf; validation data}

\item{y}{an object of class sf; SDM predictions}

\item{dens.idx}{name or index of column in \code{x} with density data;
if \code{NA}, then a column with predicted abundance
is not included in the ouput}
}
\value{
data.frame
}
\description{
Return df with overlap of presence/absence data and model
Returned df contains data columns for density, abundance, ...
...sightings (p/a)) flag, and number of animals observed
Helper function 1 for AUC, TSS, and RMSE calcs
}
