% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{eval_prediction}
\alias{eval_prediction}
\title{Returns ROCR::prediction() output of density and sightings columns
Helper function 2 for AUC and TSS calcs}
\usage{
eval_prediction(pres.pts, abs.pts, model.data, model.dens.idx,
  overlap.out = NA)
}
\description{
Returns ROCR::prediction() output of density and sightings columns
Helper function 2 for AUC and TSS calcs
}
