% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{eval_rmse}
\alias{eval_rmse}
\title{RMSE}
\usage{
eval_rmse(x, y, abund.idx, count.idx, overlap.out = NA, dens.idx = NA)
}
\arguments{
\item{x}{an object of class sf; validation data}

\item{y}{an object of class sf; SDM predictions}

\item{abund.idx}{name or index of column in \code{y} with abundance data}

\item{count.idx}{thing 1}

\item{overlap.out}{thing 2}

\item{dens.idx}{thing 3}
}
\value{
RMSE for given presence numbers and model predictions
}
\description{
RMSE
}
