% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{eval_tss}
\alias{eval_tss}
\title{Calculates TSS for given presence and absence points and model predictions
Also calculates cutoff value for predictions that maximizes sens + spec}
\usage{
eval_tss(pres.pts, abs.pts, model.data, model.dens.idx, prediction.out = NA)
}
\value{
Returns TSS value (can return vector also with cutoff value)
}
\description{
Calculates TSS for given presence and absence points and model predictions
Also calculates cutoff value for predictions that maximizes sens + spec
}
