% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{evaluation_metrics}
\alias{evaluation_metrics}
\title{Evaluation metrics}
\usage{
evaluation_metrics(x, y, x.idx, y.idx, y.idx.count = NULL)
}
\arguments{
\item{x}{an object of class sf; SDM predictions}

\item{y}{an object of class sf; validation data}

\item{x.idx}{name or index of column in \code{x} with
prediction (density) values}

\item{y.idx}{name or index of column in \code{y} with
presence/absence data}

\item{y.idx.count}{name or index of column in \code{y} with
count data. If not NULL, then RMSE is calculated}
}
\value{
A three-element, numeric vector with AUC, TSS and RMSE values, respectively.
  If \code{y.idx.count} is NULL, then the RMSE value will be NA
}
\description{
Calculate AUC, TSS, and RMSE for given presence and absence points and model predictions
}
