% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{evaluation_metrics}
\alias{evaluation_metrics}
\title{Evaluation metrics}
\usage{
evaluation_metrics(x, y, x.idx, y.idx, count.flag = FALSE)
}
\arguments{
\item{x}{object of class sf; SDM predictions}

\item{y}{object of class sf; validation data}

\item{x.idx}{name or index of column in \code{x} with prediction values}

\item{y.idx}{name or index of column in \code{y} with validation data}

\item{count.flag}{logical; if \code{TRUE} the data in column \code{y.idx} is count data,
and if \code{FALSE} the data in column \code{y.idx} is presence/absence}
}
\value{
A numeric vector with AUC, TSS and RMSE values, respectively.
  If \code{count.flag == FALSE}, the RMSE value will be \code{NA}
}
\description{
Calculate AUC, TSS, and RMSE for given density predictions and validation data
}
\details{
If \code{count.flag == TRUE}, then \code{eSDM::model_abundance(x, x.idx, FALSE)} will be run
  to calculate predicted abundance and thus calculate RMSE.
  Note that this assumes the data in column \code{x.idx} of \code{x} are density values.
  If \code{count.flag == FALSE}, then all of the values in column \code{x.idx} of \code{x} must be \code{0} or \code{1}.
  All rows of \code{x} with a value of \code{NA} in column \code{x.idx} are removed within the function
}
\examples{
evaluation_metrics(preds.1, validation.data, 1, "sight")

evaluation_metrics(preds.1, validation.data, 1, "count", TRUE)

}
