% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gshhg.l.L16}
\alias{gshhg.l.L16}
\title{Low resolution GSHHG world map}
\format{
An object of class \code{\link[sf]{sfc}}
}
\source{
\url{http://www.soest.hawaii.edu/pwessel/gshhg/}
}
\usage{
gshhg.l.L16
}
\description{
Low resolution GSHHG world map, includes hierarchical levels
  L1 and L6. Processed using \code{\link[sf:valid]{st_make_valid}}
}
\keyword{datasets}
