% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_abundance.R
\name{model_abundance}
\alias{model_abundance}
\title{Calculate predicted abundance}
\usage{
model_abundance(x, dens.idx, sum.abund = TRUE)
}
\arguments{
\item{x}{object of class \code{sf}; SDM with density predictions}

\item{dens.idx}{name or index of column(s) in \code{x} with density predictions.
Can be a character vector (column names) or numeric vector (column indices)}

\item{sum.abund}{logical; if \code{TRUE} return the total predicted abundance and
if \code{FALSE} the abundance for for each prediction polygon}
}
\value{
Depends on \code{dens.idx} and \code{sum.abund}.
  If \code{dens.idx} is a vector of length 1, then ...TODO
  either a single number (total predicted abundance for the SDM) or
  a numerical vector with the predicted abundance for each prediction polygon
}
\description{
Calculates the predicted abundance by multiplying the density predictions by prediction polygon areas
}
\details{
Multiplies the values in the specified column(s) (i.e. the density predictions)
  by the area in square kilometers of their corresponding prediciton polygon.
  The area of each prediction polygon is calcualted using \code{st_area} from \code{\link[sf]{geos_measures}}
}
\examples{
model_abundance(preds.1, "Density")
model_abundance(preds.1, c(1, 1))
model_abundance(preds.1, c(1, 1), FALSE)

}
