% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vector_processing.R
\name{model_abundance}
\alias{model_abundance}
\title{Calculate abundances for each of cols.data
Assumes that all cols.data have NAs at same place}
\usage{
model_abundance(x, dens.idx, sum.abund = TRUE)
}
\description{
Calculate abundances for each of cols.data
Assumes that all cols.data have NAs at same place
}
