% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSDM-internal.R
\name{na_pred_message}
\alias{na_pred_message}
\title{Generate message reporting length of x}
\usage{
na_pred_message(x)
}
\description{
This message was built to refer to prediction values
}
\keyword{internal}
