% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_sdm.R
\name{overlay_sdm}
\alias{overlay_sdm}
\title{Overlay SDM predictions onto base geometry}
\usage{
overlay_sdm(base.geom, sdm, overlap.perc, data.names)
}
\arguments{
\item{base.geom}{object of class sfc that sdm is being overlaid onto}

\item{sdm}{object of class sf representing the SDM that is being overlaid onto base.geom}

\item{overlap.perc}{percentage that each base polygon must be overlapped for density value to be kept}

\item{data.names}{names or indices of column(s) with data to be overlaid}
}
\description{
Overlay SDM predictions onto base geometry
}
