% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_sdm.R
\name{overlay_sdm}
\alias{overlay_sdm}
\title{Overlay SDM predictions onto base geometry}
\usage{
overlay_sdm(base.geom, sdm, overlap.perc, data.names)
}
\arguments{
\item{base.geom}{object of class \code{sfc}; base geometry}

\item{sdm}{object of class \code{sf}; original SDM predictions}

\item{overlap.perc}{percentage of each base polygon that must intersect with \code{sdm} for overlaid density value to be kept}

\item{data.names}{names or indices of column(s) with data to be overlaid}
}
\value{
Object of class \code{sf} with the geometry of \code{base.geom} and
  the specified values of \code{sdm} overalid onto that geometry
}
\description{
Overlay SDM predictions onto base geometry
}
\details{
TODO - copy from paper?
}
\examples{
overlay_sdm(sf::st_geometry(preds.1), preds.2, 50, "Density")

}
