% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_ll}
\alias{preview_ll}
\title{title}
\usage{
preview_ll(sdm.ll, data.name, title.ll, perc, col.pal, axis.cex, main.cex)
}
\arguments{
\item{sdm.ll}{sf object in crs.ll}
}
\description{
General function for plotting sf object
}
