% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_to_sfc.R
\name{pts_to_sfc_coords}
\alias{pts_to_sfc_coords}
\title{Create polygon from polygon coordinates}
\usage{
pts_to_sfc_coords(x, crs.prov = NULL)
}
\arguments{
\item{x}{data frame with two columns;
the columns must contain longitude and latitude coordinates, respectively}

\item{crs.prov}{crs value for polygon being created;
value must be readable by \code{\link{sf}{st_crs}}}
}
\value{
Object of class \code{sfc} of type \code{POLYGON} with the provided coordinates and crs value
}
\description{
Create object of class \code{sfc} of type \code{POLYGON} from a data frame with the polygon coordinates
}
\details{
TODO
}
\examples{
x <- data.frame(
  lon = c(40, 40, 50, 50, 40, NA, 20, 20, 30, 30, 20),
  lat = c(0, 10, 10, 0, 0, NA, 0, 10, 10, 0, 0)
)
pts_to_sfc_coords(x, 4326)

}
