% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_to_sfc.R
\name{pts_to_sfc_grid}
\alias{pts_to_sfc_grid}
\title{Create sfc_polygon object from data.frame with grid coordinates}
\usage{
pts_to_sfc_grid(x, poly.radius, crs.prov = NULL)
}
\arguments{
\item{x}{data.frame with at least two columns;
the columns must contain longitude and latitude coordinates, respectively,
and the coordinates must form a grid}

\item{poly.radius}{radius of each polygon to be created around points in \code{x};
must be numeric object}

\item{crs.prov}{crs value of new sfc_polygon object;
value must be interpretable by st_crs(), see st_crs() documentation}
}
\description{
Create sfc_polygon object from data.frame with only longitude and
  latitude columns, respectively, which contain the coordinates
  of the center of each polygon of a grid
}
\examples{
x <- data.frame(
  lon = c(5, 10, 15, 20, 5, 10, 15, 20),
  lat = c(rep(5, 4), rep(10, 4))
)
pts_to_sfc_grid(x, 2.5, 4326)

}
