% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validation.data}
\alias{validation.data}
\title{Sample validation data}
\format{An object of class \code{\link[sf]{sf}} with 8 rows and 3 variables
\describe{
  \item{sight}{1's and 0's indicating species presence/absence}
  \item{count}{number of individuals observed at each point}
  \item{geometry}{simple feature list column representing validation data points}
}}
\usage{
validation.data
}
\description{
Sample validation data created by cropping Validation_data.csv to the SoCal_bite.csv region
  (.csv files from ...)
}
\keyword{datasets}
